+++ packages/scene/lib/src/diff.dart
@@ -138,9 +138,7 @@ SceneDiff diffScene(SceneDocument oldDocument, SceneDocument newDocument) {
 
     final change = NodeChange(
       id,
-      transform:
-          !_transformsEqual(oldNode.transform, newNode.transform) ||
-          oldNode.excludeFromWindingParity != newNode.excludeFromWindingParity,
+      transform: !_transformsEqual(oldNode.transform, newNode.transform),
       name: oldNode.name != newNode.name,
       layers: oldNode.layers != newNode.layers,
       visible: oldNode.visible != newNode.visible,